Change frame_free_locals_no_hook to actually call* the return hook.
This change is empirical.  See Mark Williams (mwilliams) change from ~26Sep2014
  https://github.com/facebook/hhvm/commit/b104d5e960c76ea998b2bd82b067a734a4cc8f64
This is needed to get more of the agent integration tests working, since
calls to PHP's function get_keys would cause the Profiler return hook
to NOT be called for get_keys (and presumably many others), resulting
in a skew between the Profiler's notion of the stack frames (a stack of
Frame) and the actual PHP stack.